@@ -1,9 +1,9 @@
 <?php
 /**
- * @author    : JIHAD SINNAOUR
- * @package   : Apaapi | Amazon Product Advertising API Library (v5)
- * @version   : 1.1.7
- * @copyright : (c) 2019 - 2023 Jihad Sinnaour <mail@jihadsinnaour.com>
+ * @author    : Jakiboy
+ * @package   : Amazon Product Advertising API Library (v5)
+ * @version   : 1.2.0
+ * @copyright : (c) 2019 - 2024 Jihad Sinnaour <mail@jihadsinnaour.com>
  * @link      : https://jakiboy.github.io/apaapi/
  * @license   : MIT
  *
@@ -13,12 +13,13 @@
 namespace Apaapi\exceptions;
 
 /**
- * Basic Apaapi Operation Exception Class.
+ * Apaapi operation exception.
  */
 final class OperationException extends \Exception
 {
-	public static function invalidOperationRessource($ressource = null)
+	public static function invalidRessources(?string $item = null) : string
 	{
-		return "Invalid operation ressource '{$ressource}'";
+		if ( !$item ) $item = 'undefined';
+		return "Invalid operation ressources '{$item}'";
 	}
 }
