@@ -1,9 +1,9 @@
 <?php
 /**
- * @author    : JIHAD SINNAOUR
- * @package   : Apaapi | Amazon Product Advertising API Library (v5)
- * @version   : 1.1.7
- * @copyright : (c) 2019 - 2023 Jihad Sinnaour <mail@jihadsinnaour.com>
+ * @author    : Jakiboy
+ * @package   : Amazon Product Advertising API Library (v5)
+ * @version   : 1.2.0
+ * @copyright : (c) 2019 - 2024 Jihad Sinnaour <mail@jihadsinnaour.com>
  * @link      : https://jakiboy.github.io/apaapi/
  * @license   : MIT
  *
@@ -13,17 +13,18 @@
 namespace Apaapi\exceptions;
 
 /**
- * Basic Apaapi Request Exception Class.
+ * Apaapi request exception.
  */
 final class RequestException extends \Exception
 {
-	public static function invalidRequestClientMessage()
+	public static function invalidClient() : string
 	{
 		return 'Could not send request (Missing cURL and Stream functions)';
 	}
 
-	public static function invalidRequestLocaleMessage($local = null)
+	public static function invalidLocale(?string $locale = null) : string
 	{
-		return "Invalid request locale '{$local}'";
+		if ( !$locale ) $locale = 'undefined';
+		return "Invalid request locale '{$locale}'";
 	}
 }
